"Add practical example with pandas to matplotlib notebook
A real practical example with pandas was added to the matplotlib notebook to enhance understanding of data visualization. Making use of real life car sales data, series of plots were drawn to explain the various plotting methods available in matplotlib. The primary aim of these changes is to use practical data to aid understanding of data visualization."